+++ diccionario_palabras.py
@@ -1,4 +1,6 @@
 from datos import obtener_lista_definiciones
+import random
+
 
 MINIMO_CARACTERES_PALABRAS = 5
 LETRAS = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'ñ', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z'] 
@@ -64,3 +66,46 @@ def total_palabras_por_letra():
         dicc[letra] = cantidad_palabras_por_letra(letra, lista_ordenada)
 
     return dicc
+
+def obtener_letras_participantes():
+    '''
+    Crear una lista de 10 letras aleatorias de la lista de letras permitidas para el rosco.
+    '''
+    return random.sample(LETRAS, 10)
+
+
+def recibir_lista_diccionario_filtrado(diccionario_filtrado, letras_participantes):
+
+    letras_castellano = "aábcdeéfghiíjklmnñoópqrstuúvwxyz"
+    lista_palabras_participantes = []
+
+    for letra in letras_participantes:
+        palabras_candidatas = []
+
+        for item in diccionario_filtrado:
+            palabra = item[0]
+
+            palabra_sin_tildes = sacar_tildes(palabra)
+
+            if palabra_sin_tildes[0] == letra:
+                definicion = item[1]
+                # print(definicion)
+                palabras_candidatas.append([palabra, definicion])
+
+        palabra_para_esta_letra = random.choice(palabras_candidatas)
+        # print(palabra_para_esta_letra)
+        lista_palabras_participantes.append(palabra_para_esta_letra)
+
+    # sorted(lista_palabras_participantes, key=lambda i: letras_castellano.index(i[0]))
+    return sorted(lista_palabras_participantes, key=lambda i: letras_castellano.index(i[0][0]))
+
+
+# FUNCION QUE ITERA 100 VECES PARA PRUEBAS
+"""
+FUNCION QUE ITERA 100 VECES PARA PRUEBAS
+for i in range(100):
+    lista_letras2 = obtener_letras_participantes()
+    diccionario_filtrado = ordenar_filtrar_diccionario()
+    print(recibir_lista_diccionario_filtrado(diccionario_filtrado,lista_letras2))
+
+"""
